adding static linking for plugins as an option. Works about the same as with OSG. To statically link a plugin, include it in your cmake TARGET_LINK_LIBRARIES, and call USE_DTENTITYPLUGIN(pluginLibName) from a global scope. See dtEntityScript for example. This is not as sophisticated as the OSG system and I will have to do some improvements at some point, but it's a start.
I had to do a number of changes to make static linking possible. The main one: As the registering function for plugins is called from a static initializer, the ComponentPluginManager and the MessageFactory have to be globally reachable. I changed bot to be singletons. I also had to remove the member variables of the ComponentPluginManager pointing to the EntityManager, they are now passed as parameter into methods that need it.

I created a macro REGISTER_DTENTITYPLUGIN for registering plugin factories. See dtEntitySimulation/pluginfactory.cpp for usage.

The ComponentPluginFactories used to return the static TYPE member of the entity system to create. Now, ComponentPluginFactories can be created from a static initializer before that TYPE member was statically initialized. I removed the GetType virtual functin from ComponentPluginFactory and instead use the getName() method and calculate the SID at registration time.

Removed export macro usage from the plugins, they are unnecessary as only virtual methods are called on objects from these plugins